requirements.txt
Django==2.2.8
django-background-tasks==1.2.0
django-compat==1.0.15
django-crispy-forms==1.7.2
django-pagedown==1.0.6
djangorestframework==3.7.7
Markdown==3.1.1
psutil==5.6.3
python-decouple==3.1
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
